refactor token model extraction into a utility module
