+++ src/util/mod.rs
@@ -2,3 +2,4 @@
 pub mod data_helpers;
 pub mod path_helpers;
 pub mod plot;
+pub mod token_model;
